move TestConfig and SmvBaseTest into test_support package in src/main/python (#620)
This will make both available to users for client application unit tests.
