+++ src/test/python/testHive.py
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from smvbasetest import SmvBaseTest
+from test_support.smvbasetest import SmvBaseTest
 from fixture.hive.modules import M, MyHive, MyHiveWithQuery
 
 class HiveTest(SmvBaseTest):
