+++ src/test/python/testSmvFramework.py
@@ -13,7 +13,7 @@
 
 import unittest
 
-from smvbasetest import SmvBaseTest
+from test_support.smvbasetest import SmvBaseTest
 from smv import *
 from smv.dqm import *
