+++ src/test/python/testSmvGroupedData.py
@@ -13,7 +13,7 @@
 
 import unittest
 
-from smvbasetest import SmvBaseTest
+from test_support.smvbasetest import SmvBaseTest
 from smv import SmvPyCsvFile
 
 import pyspark
